arch: modify the define of uint64

+++ os/kernel/types.h
@@ -5,6 +5,6 @@ typedef unsigned char  uchar;
 typedef unsigned char uint8;
 typedef unsigned short uint16;
 typedef unsigned int  uint32;
-typedef unsigned long uint64;
+typedef unsigned long long uint64;  // RV32: long long for 64-bit support
 
-typedef uint64 pde_t;
+typedef uint32 pde_t;  // RV32 Sv32: page table entries are 32-bit
